+++ jsonpickle/backend.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import, division, unicode_literals
 
 from .compat import string_types
+from .compat import PY2
 
 
 class JSONBackend(object):
@@ -44,9 +45,9 @@ class JSONBackend(object):
         self.load_backend('ujson')
 
         # Defaults for various encoders
-        json_opts = ((), {'sort_keys': True})
+        json_opts = ((), {'sort_keys': PY2})
         self._encoder_options = {
-            'ujson': ((), {'sort_keys': True, 'escape_forward_slashes': False}),
+            'ujson': ((), {'sort_keys': PY2, 'escape_forward_slashes': False}),
             'json': json_opts,
             'simplejson': json_opts,
             'django.util.simplejson': json_opts,